PYTHON/Pantanal.dev_Arara_Azul/BackEnd+IA/requirements.txt
absl-py==1.4.0
aniso8601==9.0.1
astunparse==1.6.3
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.6.2
bs4==0.0.1
cachelib==0.9.0
cachetools==5.3.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
filelock==3.12.0
Flask==2.3.2
Flask-Caching==2.0.2
Flask-Cors==3.0.10
flask-restx==1.1.0
flatbuffers==23.5.9
fsspec==2023.5.0
gast==0.4.0
google-auth==2.18.1
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.54.2
h5py==3.8.0
huggingface-hub==0.14.1
idna==3.4
itsdangerous==2.1.2
jax==0.4.10
Jinja2==3.1.2
jsonschema==4.17.3
keras==2.12.0
libclang==16.0.0
Markdown==3.4.3
MarkupSafe==2.1.2
ml-dtypes==0.1.0
numpy==1.23.5
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.1
protobuf==4.23.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
regex==2023.5.5
requests==2.30.0
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.10.1
six==1.16.0
soupsieve==2.4.1
tensorboard==2.12.3
tensorboard-data-server==0.7.0
tensorflow==2.12.0
termcolor==2.3.0
tokenizers==0.13.3
tqdm==4.65.0
transformers==4.29.2
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
Werkzeug==2.3.4
wrapt==1.14.1
